Format gnunet.conf.5
Generally the GNU version of man formats man pages to be terrible
to read (beyond 80 characters per line), this doesn't mean we have
to make the source code of the man page(s) terrible to work in.
When a line is 400 characters long, I don't want to compare 400
characters to see the difference.

Signed-off-by: Nils Gillmann <ng0@n0.is>