Use the DNSPARSER and DNSSTUB libraries in the resolver service
We are not using the libc functions anymore for forward and backup
DNS resolutions and the DNSPARSER and DNSSTUB libraries instead.
This has the advantage that the APIs are asynchronous now and
thus multiple DNS resolutions can be done in parallel.

This breaks support for Windows and other operating systems that
don't use /etc/resolv.conf for defining DNS servers. For fixing
this the function lookup_dns_servers can be extended with different
lookup mechanisms.