RPS service: Check existence of peer before destruction
This is only a workaround.
What was causing an issue was that, during the removal of a peer (from
all datastructures), the peer was also removed from the peer_map,
causing the final destruction call to fail.
With this workaround the peer is only destroyed if it is still in the
peer_map.