gnunet-identity.1 style changes.
It is acceptable style and arguable probably more
readable for the uninitiated to see .Nd macro followed
immediat. by text instead of having the text on a new
line. The same applies to sentence length. We should
avoid very long sentences, but at the moment this seems
easier for reading and writing instead of enforcing what
the linter says (1 sentence per line), you see the same
writing style in current major users of mdoc style.