use Curve25519 for ECDH and tweetnacl where we can
This leads to some performance improvements and makes it easier to write
software that interoperates with GNUnet / GNU Taler.  It also avoids
using the rather inconvenient libgcrypt APIs.  We still need to keep
libgcrypt though, as we need it for RSA, ECDSA and some other
primitives.

This change is still behind a #define NEW_CRYPTO, as it is a breaking
change for both EdDSA (removing the superfluous additional hash) and for
ECDHE (using Curve25519 instead of Ed25519).