Improved BIO API
BIO now supports reading from and writing to in-memory buffers.

For reading, an allocated buffer (array) and a size is passed as arguments to
the function opening the handle.
For writing, a GNUNET_Buffer is created and used internally. The buffer
contents can be extracted using the relevant function.

There is a new API in addition to the existing read/write: this new API is
more "declarative" in nature and is meant to mimic APIs like GNUNET_SQ.
The read/write operations are defined in an array of specs which are then
"commited" in a single (non-atomic) operation, rather than explicitly
executing multiple function calls and checking their return value.

Also there are small changes to GNUNET_Buffer to account for BIO's new
features.

Signed-off-by: Christian Grothoff <christian@grothoff.org>