Make REQUEST_AGPL messages configurable and add handler by default
This makes two changes:

* Add a field to `struct GNUNET_OS_ProjectData' containing a URL (as a string)
pointing to the source code of the application.

* If the field is not NULL, add a handler for the REQUEST_AGPL messages
sending the specified URL to the client.

The handler is added both in client-service communications (i.e. local
services that don't make requests to other peers in the network) and in
peer-peer communications (CADET.)  This way, any client (local or remote with
CADET) can request the source code location using a standardized mechanism
instead of writing ad-hoc solutions (unless the service/peer explicitly
specifies a NULL pointer.)

Signed-off-by: Christian Grothoff <christian@grothoff.org>